Remove RecipeIngredientGroupIngredient and migrate to RecipeIngredient
All ingredients of a recipe are now defined in groups.
The title of a group is unique per recipe, there may be one empty title
RecipeIngredients that were attached to the recipe directly are now in that group.

A ingredient can only be present once in a group, if a ingredient is needed multiple times
in a recipe, it should be in different groups anyway.